backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.0.1
certifi==2026.2.25
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.19.2
exceptiongroup==1.3.1
fastapi==0.110.0
fastapi-cli==0.0.18
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.6.3
pycparser==2.23
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
Pygments==2.19.2
PyMySQL==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.36.3
tomli==2.4.1
typer==0.20.1
typing_extensions==4.13.2
uvicorn==0.27.1
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
